test: split oversized agent resolve config tests
Move validate and normalize/role cases out of agent_resolve_test.go so
each file stays under the AGENTS.md 300-line subsystem test guidance.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
